chore(1-3421): move release plan component to legacy release plan
